+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SphericalSurface3D: line_intersections, linesegment_intersections.
 - Sweep with muitiform profile contour.
 - OpenShell3D: from_faces (using faces graph)
+- SphericalFace3D: from_contours3d_and_rectangular_cut
 
 ### Fixed
 - ClosedShell3D: is_face_inside, get_subtraction_valid_faces, valid_intersection_faces, point_belongs
@@ -122,6 +123,7 @@ All adequations have been done for all tests and existing scripts.
 - DisplayMesh3D: triangulation_faces
 - Woodpecker CI setup
 - ContourMixin: primitive_section_over_contour.
+- Face3D: split_by_plane
 
 ### Fixed
 - 2D conversion: create 2D function name in core_compiled
@@ -218,6 +220,7 @@ All adequations have been done for all tests and existing scripts.
 - PlaneFace3D: merge_faces
 - Contour2D: divide.
 - BSplineFace3D: test_linesegment_intersections_approximation.
+- CylindricalFace3D: split_by_plane.
 
 v0.10.0 [Released 20/04/2023]